requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
deep-translator==1.11.4
distro==1.9.0
ffmpeg-python==0.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
moviepy==1.0.3
multidict==6.0.5
mutagen==1.47.0
mysql-connector-python==8.4.0
numpy==2.0.1
openai==1.31.0
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
priority==2.0.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
pytube==15.0.0
Quart==0.19.6
requests==2.32.3
rsa==4.9
selenium==4.23.1
setuptools==72.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
stripe==9.9.0
tqdm==4.66.4
trio==0.26.2
trio-websocket==0.11.1
twilio==9.1.0
typing_extensions==4.12.1
urllib3==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
yarl==1.9.4
yt-dlp==2024.8.1

